+++ tests/upgrade.rs
@@ -1,19 +1,19 @@
+mod common;
+
 mod upgrade {
-	use http::{header, HeaderMap, StatusCode};
-	use reqwest::Client;
+	use crate::common::docker::DockerContainer;
+	use crate::common::expected::Expected;
+	use crate::common::rest_client::RestClient;
 	use serde_json::Value as JsonValue;
-	use std::process::Command;
-	use std::time::{Duration, SystemTime};
+	use std::time::Duration;
 	use surrealdb::engine::any::{connect, Any};
 	use surrealdb::{Connection, Surreal};
 	use test_log::test;
-	use tokio::time::sleep;
-	use tracing::{debug, error, info, warn};
+	use tracing::info;
 	use ulid::Ulid;
 
 	const PREVIOUS_DOCKER_VERSION: &str = "SURREALDB_TEST_DOCKER_PREVIOUS_VERSION";
 	const DEFAULT_DOCKER_VERSION: &str = "v1.2.1";
-	const DOCKER_EXPOSED_PORT: usize = 8000;
 	const CNX_TIMEOUT: Duration = Duration::from_secs(180);
 	const NS: &str = "test";
 	const DB: &str = "test";
@@ -38,11 +38,14 @@ mod upgrade {
 
 		{
 			// Start the docker instance
-			let mut docker = DockerContainer::start(&docker_version, &file_path);
-			let client = RestClient::new().wait_for_connection().await.unwrap_or_else(|| {
-				docker.logs();
-				panic!("No connected client")
-			});
+			let mut docker = DockerContainer::start(&docker_version, &file_path, USER, PASS);
+			let client = RestClient::new(&NS, &DB, &USER, &PASS)
+				.wait_for_connection(&CNX_TIMEOUT)
+				.await
+				.unwrap_or_else(|| {
+					docker.logs();
+					panic!("No connected client")
+				});
 			// Create data samples
 			if docker_version.starts_with("1.0.") {
 				create_data_for_1_0(&client).await;
@@ -76,12 +79,12 @@ mod upgrade {
 
 	// Set of DATA for Full Text Search
 	const DATA_FTS: [&str; 5] = [
-        "DEFINE ANALYZER name TOKENIZERS class FILTERS lowercase,ngram(1,128)",
-        "DEFINE ANALYZER userdefinedid TOKENIZERS blank FILTERS lowercase,ngram(1,32)",
-        "DEFINE INDEX account_name_search_idx ON TABLE account COLUMNS name SEARCH ANALYZER name BM25(1.2,0.75) HIGHLIGHTS",
-        "DEFINE INDEX account_user_defined_id_search_idx ON TABLE account COLUMNS user_defined_id SEARCH ANALYZER userdefinedid BM25 HIGHLIGHTS",
-        "CREATE account SET name='Tobie', user_defined_id='Tobie'",
-    ];
+		"DEFINE ANALYZER name TOKENIZERS class FILTERS lowercase,ngram(1,128)",
+		"DEFINE ANALYZER userdefinedid TOKENIZERS blank FILTERS lowercase,ngram(1,32)",
+		"DEFINE INDEX account_name_search_idx ON TABLE account COLUMNS name SEARCH ANALYZER name BM25(1.2,0.75) HIGHLIGHTS",
+		"DEFINE INDEX account_user_defined_id_search_idx ON TABLE account COLUMNS user_defined_id SEARCH ANALYZER userdefinedid BM25 HIGHLIGHTS",
+		"CREATE account SET name='Tobie', user_defined_id='Tobie'",
+	];
 
 	// Set of QUERY and RESULT to check for Full Text Search
 	static CHECK_FTS: [Check; 1] =
@@ -108,33 +111,6 @@ mod upgrade {
 		 Expected::Two("{\"dist\": 2.0, \"id\": {\"tb\": \"pts\", \"id\": {\"Number\": 1}}}", "{ \"dist\": 4.0, \"id\": {\"tb\": \"pts\", \"id\": {\"Number\": 2}}}"))];
 
 	type Check = (&'static str, Expected);
-	enum Expected {
-		Any,
-		One(&'static str),
-		Two(&'static str, &'static str),
-	}
-
-	impl Expected {
-		fn check_results(&self, q: &str, results: &[JsonValue]) {
-			match self {
-				Expected::Any => {}
-				Expected::One(expected) => {
-					assert_eq!(results.len(), 1, "Wrong number of result for {}", q);
-					Self::check_json(q, &results[0], expected);
-				}
-				Expected::Two(expected1, expected2) => {
-					assert_eq!(results.len(), 2, "Wrong number of result for {}", q);
-					Self::check_json(q, &results[0], expected1);
-					Self::check_json(q, &results[1], expected2);
-				}
-			}
-		}
-
-		fn check_json(q: &str, result: &JsonValue, expected: &str) {
-			let expected: JsonValue = serde_json::from_str(expected).expect(expected);
-			assert_eq!(result, &expected, "Unexpected result on query {}", q);
-		}
-	}
 
 	const CHECK_DB: [Check; 1] = [("INFO FOR DB", Expected::Any)];
 
@@ -206,176 +182,5 @@ mod upgrade {
 		db.use_ns(NS).await.unwrap();
 		db.use_db(DB).await.unwrap();
 		db
-	}
-
-	struct DockerContainer {
-		id: String,
-		running: bool,
-	}
-
-	impl DockerContainer {
-		fn start(version: &str, file_path: &str) -> Self {
-			let docker_image = format!("surrealdb/surrealdb:{version}");
-			info!("Start Docker image {docker_image} with file {file_path}");
-			let mut args = Arguments::new([
-				"run",
-				"-p",
-				&format!("127.0.0.1:8000:{DOCKER_EXPOSED_PORT}"),
-				"-d",
-			]);
-			args.add([docker_image]);
-			args.add(["start", "--auth", "--user", USER, "--pass", PASS]);
-			args.add([format!("file:{file_path}")]);
-			let id = Self::docker(args);
-			Self {
-				id,
-				running: true,
-			}
-		}
-
-		fn logs(&self) {
-			info!("Logging Docker container {}", self.id);
-			Self::docker(Arguments::new(["logs", &self.id]));
-		}
-		fn stop(&mut self) {
-			if self.running {
-				info!("Stopping Docker container {}", self.id);
-				Self::docker(Arguments::new(["stop", &self.id]));
-				self.running = false;
-			}
-		}
-
-		fn extract_data_dir(&self, file_path: &str) {
-			let container_src_path = format!("{}:{file_path}", self.id);
-			info!("Extract directory from Docker container {}", container_src_path);
-			Self::docker(Arguments::new(["cp", &container_src_path, file_path]));
-		}
-
-		fn docker(args: Arguments) -> String {
-			let mut command = Command::new("docker");
-
-			let output = command.args(args.0).output().unwrap();
-			let std_out = String::from_utf8(output.stdout).unwrap().trim().to_string();
-			if !output.stderr.is_empty() {
-				error!("{}", String::from_utf8(output.stderr).unwrap());
-			}
-			assert_eq!(output.status.code(), Some(0), "Docker command failure: {:?}", command);
-			std_out
-		}
-	}
-
-	impl Drop for DockerContainer {
-		fn drop(&mut self) {
-			// Be sure the container is stopped
-			self.stop();
-			// Delete the container
-			info!("Delete Docker container {}", self.id);
-			Self::docker(Arguments::new(["rm", &self.id]));
-		}
-	}
-
-	struct Arguments(Vec<String>);
-
-	impl Arguments {
-		fn new<I, S>(args: I) -> Self
-		where
-			I: IntoIterator<Item = S>,
-			S: Into<String>,
-		{
-			let mut a = Self(vec![]);
-			a.add(args);
-			a
-		}
-
-		fn add<I, S>(&mut self, args: I)
-		where
-			I: IntoIterator<Item = S>,
-			S: Into<String>,
-		{
-			for arg in args {
-				self.0.push(arg.into());
-			}
-		}
-	}
-
-	struct RestClient {
-		c: Client,
-		u: String,
-	}
-
-	impl RestClient {
-		fn new() -> Self {
-			let mut headers = HeaderMap::new();
-			headers.insert("NS", NS.parse().unwrap());
-			headers.insert("DB", DB.parse().unwrap());
-			headers.insert(header::ACCEPT, "application/json".parse().unwrap());
-			let c = Client::builder()
-				.connect_timeout(Duration::from_millis(10))
-				.default_headers(headers)
-				.build()
-				.expect("Client::builder()...build()");
-			Self {
-				c,
-				u: format!("http://127.0.0.1:{DOCKER_EXPOSED_PORT}/sql"),
-			}
-		}
-
-		async fn wait_for_connection(self) -> Option<Self> {
-			sleep(Duration::from_secs(2)).await;
-			let start = SystemTime::now();
-			while start.elapsed().unwrap() < CNX_TIMEOUT {
-				sleep(Duration::from_secs(2)).await;
-				if let Some(r) = self.query("INFO FOR ROOT").await {
-					if r.status() == StatusCode::OK {
-						return Some(self);
-					}
-				}
-				warn!("DB not yet responding");
-				sleep(Duration::from_secs(2)).await;
-			}
-			None
-		}
-
-		async fn query(&self, q: &str) -> Option<reqwest::Response> {
-			match self.c.post(&self.u).basic_auth(USER, Some(PASS)).body(q.to_string()).send().await
-			{
-				Ok(r) => Some(r),
-				Err(e) => {
-					error!("{e}");
-					None
-				}
-			}
-		}
-
-		async fn checked_query(&self, q: &str, expected: &Expected) {
-			let r = self.query(q).await.unwrap_or_else(|| panic!("No response for {q}"));
-			assert_eq!(
-				r.status(),
-				StatusCode::OK,
-				"Wrong response for {q} -> {}",
-				r.text().await.expect(q)
-			);
-			// Convert the result to JSON
-			let j: JsonValue = r.json().await.expect(q);
-			debug!("{q} => {j:#}");
-			// The result should be an array
-			let results_with_status = j.as_array().expect(q);
-			assert_eq!(results_with_status.len(), 1, "Wrong number of results on query {q}");
-			let result_with_status = &results_with_status[0];
-			// Check the status
-			let status = result_with_status
-				.get("status")
-				.unwrap_or_else(|| panic!("No status on query: {q}"));
-			assert_eq!(status.as_str(), Some("OK"), "Wrong status for {q} => {status:#}");
-			// Extract the results
-			let results = result_with_status
-				.get("result")
-				.unwrap_or_else(|| panic!("No result for query: {q}"));
-			if !matches!(expected, Expected::Any) {
-				// Check the results
-				let results = results.as_array().expect(q);
-				expected.check_results(q, results);
-			}
-		}
 	}
 }
